+++ HISTORY.md
@@ -17,6 +17,7 @@
 * extensions: added "result of process" extension
 * extensions: added "object of process" extension
 * extensions, threads: added "step process" extension
+* blocks, objects, tables: added "step" entry to paused process context menu
 
 ### 2026-07-30
 * symbols: new animated "broken gears" symbols